fix file name and improve the diagram
